apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seensnap-api"
version = "0.1.0"
description = "SeenSnap FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116,<1.0",
  "uvicorn[standard]>=0.35,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.16,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.10,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "httpx>=0.28,<1.0",
  "google-auth>=2.40,<3.0",
  "PyJWT>=2.10,<3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4,<9.0",
  "pytest-asyncio>=1.1,<2.0",
  "ruff>=0.12,<1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
